backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pandas==2.1.3
python-dotenv==1.0.0
httpx==0.25.1
openai==1.3.5
python-multipart==0.0.6
